requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.1
discord.py==2.0.0
et-xmlfile==1.1.0
Flask==2.2.2
frozenlist==1.3.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
openpyxl==3.0.10
pip==22.1.2
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
replit==3.2.4
requests==2.28.1
setuptools==63.4.1
soupsieve==2.3.2.post1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
yarl==1.8.1
